+++ tests/testthat/helper-fixtures.R
@@ -29,3 +29,7 @@ test_fixtures$half_empty_gql_response <- list(
     )
   )
 )
+
+test_fixtures$gl_search_response <- list(
+
+)