+++ tests/testthat/test-01-GQLQueryGitLab.R
@@ -18,10 +18,15 @@ test_that("user query is built properly", {
   test_mocker$cache(gl_user_query)
 })
 
-test_that("file query is built properly", {
+test_that("file queries are built properly", {
   gl_files_query <-
     test_gqlquery_gl$files_by_org()
   expect_snapshot(
     gl_files_query
+  )
+  gl_repo_files_query <-
+    test_gqlquery_gl$files_from_repo()
+  expect_snapshot(
+    gl_repo_files_query
   )
 })