+++ tests/testthat/test-03-EngineGraphQLGitHub.R
@@ -25,7 +25,7 @@ test_that("`pull_commits_page_from_repo()` pulls commits page from repository", 
     date_from = "2023-01-01",
     date_until = "2023-02-28"
   )
-  expect_gh_commit_gql(
+  expect_gh_commit_gql_response(
     commits_page
   )
   test_mocker$cache(commits_page)
@@ -41,7 +41,7 @@ test_that("`pull_repos_page_from_org()` pulls repos page from GitHub organizatio
     from = "org",
     org = "r-world-devs"
   )
-  expect_gh_repos(
+  expect_gh_repos_gql_response(
     gh_repos_page
   )
   test_mocker$cache(gh_repos_page)
@@ -78,7 +78,7 @@ test_that("`pull_repos_page()` pulls repos page from GitHub user", {
     from = "user",
     user = "maciekbanas"
   )
-  expect_gh_user_repos(
+  expect_gh_user_repos_gql_response(
     gh_repos_user_page
   )
   test_mocker$cache(gh_repos_user_page)
@@ -200,8 +200,9 @@ test_that("GitHub prepares user table", {
   gh_user_table <- test_gql_gh$prepare_user_table(
     user_response = test_mocker$use("gh_user_response")
   )
-  expect_user_table(
-    gh_user_table
+  expect_users_table(
+    gh_user_table,
+    one_user = TRUE
   )
   test_mocker$cache(gh_user_table)
 })
@@ -239,11 +240,10 @@ test_that("`pull_repos()` works as expected", {
     "private$pull_repos",
     test_mocker$use("gh_repos_from_org")
   )
-  settings <- list(search_param = "org")
   expect_snapshot(
     gh_repos_org <- test_gql_gh$pull_repos(
       org = "r-world-devs",
-      settings = settings
+      settings = test_settings
     )
   )
   expect_repos_table(
@@ -255,14 +255,13 @@ test_that("`pull_repos()` works as expected", {
     "private$pull_repos_from_team",
     test_mocker$use("gh_repos_from_team")
   )
-  settings <- list(
-    search_param = "team",
-    team = test_team
-  )
+  test_settings[["search_param"]] <- "team"
+  test_settings[["team"]] <- test_team
+
   expect_snapshot(
     gh_repos_team <- test_gql_gh$pull_repos(
       org = "r-world-devs",
-      settings = settings
+      settings = test_settings
     )
   )
   expect_repos_table(
@@ -284,7 +283,7 @@ test_that("`pull_commits()` retrieves commits in the table format", {
   )
 
   repos_table <- test_mocker$use("gh_repos_table") %>%
-    dplyr::filter(name == "GitStats")
+    dplyr::filter(repo_name == "GitStats")
 
   mockery::stub(
     test_gql_gh$pull_commits,
@@ -292,14 +291,12 @@ test_that("`pull_commits()` retrieves commits in the table format", {
     repos_table
   )
 
-  settings <- list(search_param = "org")
-
   expect_snapshot(
     commits_table <- test_gql_gh$pull_commits(
       org = "r-world-devs",
       date_from = "2023-01-01",
       date_until = "2023-02-28",
-      settings = settings
+      settings = test_settings
     )
   )